+++ factory/declarations.py
@@ -22,6 +22,7 @@
 
 
 import itertools
+import warnings
 
 from . import compat
 from . import utils
@@ -428,9 +429,18 @@ class RelatedFactory(PostGenerationDeclaration):
             calling the related factory
     """
 
-    def __init__(self, factory, name='', **defaults):
+    def __init__(self, factory, factory_related_name='', **defaults):
         super(RelatedFactory, self).__init__()
-        self.name = name
+        if factory_related_name == '' and defaults.get('name') is not None:
+            warnings.warn(
+                "Usage of RelatedFactory(SomeFactory, name='foo') is deprecated"
+                " and will be removed in the future. Please use the"
+                " RelatedFactory(SomeFactory, 'foo') or"
+                " RelatedFactory(SomeFactory, factory_related_name='foo')"
+                " syntax instead", PendingDeprecationWarning, 2)
+            factory_related_name = defaults.pop('name')
+
+        self.name = factory_related_name
         self.defaults = defaults
 
         if isinstance(factory, type):